# app/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "jazz-life",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/auth/callback": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Callback",
        "operationId": "callback_api_auth_callback_get",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "error",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/artists": {
      "get": {
        "tags": [
          "artists"
        ],
        "summary": "List Artists",
        "operationId": "list_artists_api_artists_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse_ArtistRead_"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "artists"
        ],
        "summary": "Upsert Artist",
        "description": "Upsert keyed on spotify_id; returns the existing row if present (200).\n\nPhase B-3 PR-2: called by RecordFormModal when a Spotify album is selected\nand the album's artist is not yet in the DB. Idempotent by design: a\nduplicate POST never returns 409.",
        "operationId": "upsert_artist_api_artists_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtistCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/artists/{spotify_id}": {
      "get": {
        "tags": [
          "artists"
        ],
        "summary": "Get Artist",
        "description": "Fetch a single artist, filling in image_url from Spotify when it is NULL.\n\nThe \"lazy photo hydration\" endpoint called when ArtistDetailModal opens.\nSpotify-side errors are swallowed in the service layer, so when no image\nis available the artist is returned with image_url=null (the UI falls back\nto initials).",
        "operationId": "get_artist_api_artists__spotify_id__get",
        "parameters": [
          {
            "name": "spotify_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Spotify Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/records": {
      "get": {
        "tags": [
          "records"
        ],
        "summary": "List Records",
        "description": "Return the current user's collection (ADR-006 §3.4).\n\nPassing `limit` paginates. The response `total` is the size of the user's\ncollection (after the `status` filter), so the frontend can derive the page\ncount with `Math.ceil(total / limit)`.\n\n- `status`: filter by owned / wanted (the ADR-013 Hunt list uses `wanted`).\n- `sort`: `artist` (artist name, then title, ascending) / `added`\n  (created_at descending). Defaults to\n  `is_pinned DESC, pin_order ASC, display_order ASC` (for the Home matrix).",
        "operationId": "list_records_api_records_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "owned",
                    "wanted"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "artist",
                    "added"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sort"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse_VinylRecordRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "records"
        ],
        "summary": "Create Record",
        "description": "Create one record: catalog find-or-create + user_collections INSERT +\nauto-follow in a single transaction. UNIQUE(user_id, vinyl_record_id)\nviolations return 409.",
        "operationId": "create_record_api_records_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VinylRecordCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VinylRecordRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/records/pins/order": {
      "put": {
        "tags": [
          "records"
        ],
        "summary": "Reorder Pins",
        "description": "Persist the order of pinned records after a drag & drop.\n\n`ids` in the body is every currently pinned user_collection.id in the\ndesired order. 409 if it does not match the current pin set.",
        "operationId": "reorder_pins_api_records_pins_order_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PinReorderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/records/{id}": {
      "put": {
        "tags": [
          "records"
        ],
        "summary": "Update Record",
        "description": "Partially update user_collections, guarded by user_id. Catalog fields\nare written only when `source='manual'` (ADR-006 §2.5). Setting\n`spotify_album_id` takes the manual→spotify promotion path (§2.9).",
        "operationId": "update_record_api_records__id__put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VinylRecordUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VinylRecordRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "records"
        ],
        "summary": "Delete Record",
        "description": "Hard-delete the user_collections row (favorites CASCADE). The catalog is untouched.",
        "operationId": "delete_record_api_records__id__delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ArtistCreate": {
        "properties": {
          "spotify_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Spotify Id"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "source": {
            "type": "string",
            "enum": [
              "seeded",
              "spotify_dynamic",
              "manual"
            ],
            "title": "Source",
            "default": "spotify_dynamic"
          }
        },
        "type": "object",
        "required": [
          "spotify_id",
          "name"
        ],
        "title": "ArtistCreate",
        "description": "Create-artist request; upserts (returns the existing row if spotify_id exists).\n\nCalled by the Phase B-3 PR-2 RecordFormModal when a Spotify album is\nselected and the album's artist is not yet in the DB."
      },
      "ArtistRead": {
        "properties": {
          "spotify_id": {
            "type": "string",
            "title": "Spotify Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "source": {
            "type": "string",
            "enum": [
              "seeded",
              "spotify_dynamic",
              "manual"
            ],
            "title": "Source"
          },
          "added_at": {
            "type": "string",
            "format": "date-time",
            "title": "Added At"
          }
        },
        "type": "object",
        "required": [
          "spotify_id",
          "name",
          "image_url",
          "source",
          "added_at"
        ],
        "title": "ArtistRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ListResponse_ArtistRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ArtistRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ListResponse[ArtistRead]"
      },
...
      "ListResponse_VinylRecordRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/VinylRecordRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ListResponse[VinylRecordRead]"
      },
      "PinReorderRequest": {
        "properties": {
          "ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Ids"
          }
        },
        "type": "object",
        "required": [
          "ids"
        ],
        "title": "PinReorderRequest",
        "description": "The id sequence of pinned records after a drag & drop reorder.\n\n`ids` is every row the user currently has pinned, in the user's desired\norder. The backend renumbers `pin_order` as 1..N. Missing / duplicate /\nunpinned ids return 409."
      },
...
      "VinylRecordCreate": {
        "properties": {
          "artist_id": {
            "type": "string",
            "title": "Artist Id"
          },
          "spotify_album_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spotify Album Id"
          },
          "source": {
            "type": "string",
            "enum": [
              "spotify",
              "manual"
            ],
            "title": "Source",
            "default": "manual"
          },
          "status": {
            "type": "string",
            "enum": [
              "owned",
              "wanted"
            ],
            "title": "Status",
            "default": "owned"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "original_release_date": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\d{4}(-\\d{2}(-\\d{2})?)?$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Release Date"
          },
          "pressing_info": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pressing Info"
          },
          "purchase_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purchase Date"
          },
          "purchase_store": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purchase Store"
          },
          "purchase_price": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purchase Price"
          },
          "purchase_currency": {
            "type": "string",
            "pattern": "^[A-Z]{3}$",
            "title": "Purchase Currency",
            "default": "JPY"
          },
          "rating": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 5.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "memo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memo"
          },
          "favorite_tracks": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FavoriteTrack"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Favorite Tracks"
          }
        },
        "type": "object",
        "required": [
          "artist_id",
          "title"
        ],
        "title": "VinylRecordCreate"
      },
      "VinylRecordRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "artist_id": {
            "type": "string",
            "title": "Artist Id"
          },
          "spotify_album_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spotify Album Id"
          },
          "source": {
            "type": "string",
            "enum": [
              "spotify",
              "manual"
            ],
            "title": "Source"
          },
          "status": {
            "type": "string",
            "enum": [
              "owned",
              "wanted"
            ],
            "title": "Status"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "original_release_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Release Date"
          },
          "pressing_info": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pressing Info"
          },
          "purchase_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purchase Date"
          },
          "purchase_store": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purchase Store"
          },
          "purchase_price": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purchase Price"
          },
          "purchase_currency": {
            "type": "string",
            "title": "Purchase Currency"
          },
          "rating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "memo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memo"
          },
          "favorite_tracks": {
            "items": {
              "$ref": "#/components/schemas/FavoriteTrack"
            },
            "type": "array",
            "title": "Favorite Tracks"
          },
          "display_order": {
            "type": "integer",
            "title": "Display Order"
          },
          "is_pinned": {
            "type": "boolean",
            "title": "Is Pinned"
          },
          "pin_order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pin Order"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "artist_id",
          "spotify_album_id",
          "source",
          "status",
          "title",
          "image_url",
          "original_release_date",
          "pressing_info",
          "purchase_date",
          "purchase_store",
          "purchase_price",
          "purchase_currency",
          "rating",
          "memo",
          "favorite_tracks",
          "display_order",
          "is_pinned",
          "pin_order",
          "created_at",
          "updated_at"
        ],
        "title": "VinylRecordRead",
        "description": "ADR-006: the response shape stays flat, combining catalog + ownership.\n\n`id` is `user_collections.id` (the frontend sees it as a single record).\n`favorite_tracks` is a structured list (ascending by position)."
      },
...
    }
  }
}